Rename person package into contact
