+++ src/main/java/seedu/address/model/util/SampleDataUtil.java
@@ -6,9 +6,9 @@ import java.util.stream.Collectors;
 
 import seedu.address.model.ModuleList;
 import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
-import seedu.address.model.person.Contact;
+import seedu.address.model.contact.Email;
+import seedu.address.model.contact.Name;
+import seedu.address.model.contact.Contact;
 import seedu.address.model.tag.Tag;
 
 /**
