+++ src/test/java/seedu/address/logic/LogicManagerTest.java
@@ -25,7 +25,7 @@ import seedu.address.model.ModelManager;
 //import seedu.address.model.ReadOnlyAddressBook;
 // import seedu.address.model.UserPrefs;
 import seedu.address.model.ReadOnlyModuleList;
-import seedu.address.model.person.Contact;
+import seedu.address.model.contact.Contact;
 import seedu.address.storage.JsonModuleListStorage;
 import seedu.address.storage.JsonUserPrefsStorage;
 import seedu.address.storage.StorageManager;
