+++ src/test/java/seedu/address/logic/parser/AddCommandParserTest.java
@@ -28,7 +28,7 @@ import org.junit.jupiter.api.Test;
 import seedu.address.logic.commands.AddCommand;
 //import seedu.address.model.person.Email;
 //import seedu.address.model.person.Name;
-import seedu.address.model.person.Contact;
+import seedu.address.model.contact.Contact;
 //import seedu.address.model.tag.Tag;
 import seedu.address.testutil.PersonBuilder;
