+++ src/test/java/seedu/address/model/contact/AddressTest.java
@@ -1,0 +1,6 @@
+package seedu.address.model.contact;
+
+public class AddressTest {
+
+
+}
