+++ src/test/java/seedu/address/model/contact/PhoneTest.java
@@ -1,0 +1,4 @@
+package seedu.address.model.contact;
+
+public class PhoneTest {
+}
