--- src/test/java/seedu/address/model/person/AddressTest.java
@@ -1,6 +1,0 @@
-package seedu.address.model.person;
-
-public class AddressTest {
-
-
-}
