--- src/test/java/seedu/address/model/person/PhoneTest.java
@@ -1,4 +1,0 @@
-package seedu.address.model.person;
-
-public class PhoneTest {
-}
