+++ src/test/java/seedu/address/testutil/PersonUtil.java
@@ -7,7 +7,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
 //import java.util.Set;
 import seedu.address.logic.commands.AddCommand;
 import seedu.address.logic.commands.EditCommand;
-import seedu.address.model.person.Contact;
+import seedu.address.model.contact.Contact;
 //import seedu.address.model.tag.Tag;
 
 /**
