+++ src/test/java/seedu/address/testutil/TypicalPersons.java
@@ -12,7 +12,7 @@ import java.util.Arrays;
 import java.util.List;
 
 import seedu.address.model.ModuleList;
-import seedu.address.model.person.Contact;
+import seedu.address.model.contact.Contact;
 
 /**
  * A utility class containing a list of {@code Person} objects to be used in tests.
